requirements.txt
requests==2.31.0
pandas==2.0.3
prophet==1.1.5
sqlalchemy==2.0.23
python-dotenv==1.0.0
fredapi==0.5.1
schedule==1.2.0